Webapp/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
alabaster=0.7.12=py37_0
anaconda=custom=py37_0
anaconda-client=1.7.2=py37_0
anaconda-project=0.8.2=py37_0
asn1crypto=0.24.0=py37_0
astroid=2.2.5=py37_0
astropy=3.1.2=py37h7b6447c_0
atomicwrites=1.3.0=py37_1
attrs=19.1.0=py37_1
azure-common=1.1.22=py_0
azure-nspkg=3.0.2=py_0
azure-storage-blob=1.5.0=py_0
azure-storage-common=1.4.1=py_0
babel=2.6.0=py37_0
backcall=0.1.0=py37_0
backports=1.0=py37_1
backports.os=0.1.1=py37_0
backports.shutil_get_terminal_size=1.0.0=py37_2
beautifulsoup4=4.7.1=py37_1
bitarray=0.8.3=py37h14c3975_0
bkcharts=0.2=py37_0
blas=2.7=openblas
bleach=3.1.0=py37_0
blosc=1.15.0=hd408876_0
bokeh=1.0.4=py37_0
boto=2.49.0=py37_0
bottleneck=1.2.1=py37h035aef0_1
bzip2=1.0.6=h14c3975_5
ca-certificates=2019.6.16=hecc5488_0
cairo=1.16.0=h18b612c_1001
certifi=2019.6.16=py37_0
cffi=1.12.2=py37h2e261b9_1
chardet=3.0.4=py37_1
click=7.0=py37_0
cloudpickle=0.8.0=py37_0
clyent=1.2.2=py37_1
colorama=0.4.1=py37_0
contextlib2=0.5.5=py37_0
cryptography=2.6.1=py37h1ba5d50_0
curl=7.64.0=hbc83047_2
cycler=0.10.0=py37_0
cython=0.29.6=py37he6710b0_0
cytoolz=0.9.0.1=py37h14c3975_1
dask=1.1.4=py37_1
dask-core=1.1.4=py37_1
dbus=1.13.6=he372182_0
decorator=4.4.0=py37_1
defusedxml=0.5.0=py37_1
distributed=1.26.0=py37_1
docutils=0.14=py37_0
entrypoints=0.3=py37_0
et_xmlfile=1.0.1=py37_0
expat=2.2.6=he6710b0_0
fastcache=1.0.2=py37h14c3975_2
ffmpeg=4.1.3=h167e202_0
filelock=3.0.10=py37_0
flask=1.0.2=py37_1
fontconfig=2.13.1=he4413a7_1000
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
get_terminal_size=1.0.0=haa9412d_0
gettext=0.19.8.1=hc5be6a0_1002
gevent=1.4.0=py37h7b6447c_0
giflib=5.1.9=h516909a_0
glib=2.58.3=h6f030ca_1001
glob2=0.6=py37_1
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py37h10f8cd9_2
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=h23475e2_0
greenlet=0.4.15=py37h7b6447c_0
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
h5py=2.9.0=py37h7918eee_0
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.4=hb1b8bf9_0
heapdict=1.0.0=py37_2
html5lib=1.0.1=py37_0
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
imageio=2.5.0=py37_0
imagesize=1.1.0=py37_0
importlib_metadata=0.8=py37_0
intel-openmp=2019.3=199
ipykernel=5.1.0=py37h39e3cac_0
ipython=7.4.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.4.2=py37_0
isort=4.3.16=py37_0
itsdangerous=1.1.0=py37_0
jasper=1.900.1=h07fcdf6_1006
jbig=2.1=hdba287a_0
jdcal=1.4=py37_0
jedi=0.13.3=py37_0
jeepney=0.4=py37_0
jinja2=2.10=py37_0
joblib=0.13.2=py_0
jpeg=9c=h14c3975_1001
jsonschema=3.0.1=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.2.4=py37_0
jupyter_console=6.0.0=py37_0
jupyter_core=4.4.0=py37_0
jupyterlab=0.35.4=py37hf63ae98_0
jupyterlab_server=0.2.0=py37_0
keyring=18.0.0=py37_0
kiwisolver=1.0.1=py37hf484d3e_0
krb5=1.16.1=h173b8e3_7
lame=3.100=h14c3975_1001
lazy-object-proxy=1.3.1=py37h14c3975_2
libarchive=3.3.3=h5d8350f_5
libblas=3.8.0=7_openblas
libcblas=3.8.0=7_openblas
libcurl=7.64.0=h20c2e04_2
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=7_openblas
liblapacke=3.8.0=7_openblas
liblief=0.9.0=h7725739_2
libopenblas=0.3.6=h5a2b251_0
libpng=1.6.36=hbc83047_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.0=h1ba5d50_4
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libtool=2.4.6=h7b6447c_5
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
libxslt=1.1.33=h7d1a2b0_0
llvmlite=0.28.0=py37hd408876_0
locket=0.2.0=py37_1
lxml=4.3.2=py37hefd8a0e_0
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h49e0be7_2
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.0.3=py37h5429711_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
mkl=2019.3=199
mkl-service=2.0.2=py37h516909a_0
mkl_fft=1.0.13=py37h516909a_1
mkl_random=1.0.4=py37hf2d7682_0
more-itertools=6.0.0=py37_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py37_0
msgpack-python=0.6.1=py37hfd86e86_1
multipledispatch=0.6.0=py37_0
nbconvert=5.4.1=py37_3
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.2=py37_1
nltk=3.4=py37_1
nose=1.3.7=py37_2
notebook=5.7.8=py37_0
numba=0.43.1=py37h962f231_0
numexpr=2.6.9=py37h637b7d7_1000
numpy=1.16.4=py37h99e49ec_0
numpy-base=1.16.4=py37h2f8d375_0
numpydoc=0.8.0=py37_0
olefile=0.46=py37_0
openblas=0.3.5=h9ac9557_1001
opencv=4.1.0=py37h5517eff_4
openh264=1.8.0=hdbcaa40_1000
openpyxl=2.6.1=py37_1
openssl=1.1.1b=h14c3975_1
packaging=19.0=py37_0
pandas=0.24.2=py37he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
pango=1.40.14=h4ea9474_1004
parso=0.3.4=py37_0
partd=0.3.10=py37_1
patchelf=0.9=he6710b0_3
path.py=11.5.0=py37_0
pathlib2=2.3.3=py37_0
patsy=0.5.1=py37_0
pcre=8.43=he6710b0_0
pep8=1.7.1=py37_0
pexpect=4.6.0=py37_0
pickleshare=0.7.5=py37_0
pillow=5.4.1=py37h34e0f95_0
pip=19.0.3=py37_0
pixman=0.38.0=h7b6447c_0
pkginfo=1.5.0.1=py37_0
pluggy=0.9.0=py37_0
ply=3.11=py37_0
prometheus_client=0.6.0=py37_0
prompt_toolkit=2.0.9=py37_0
psutil=5.6.1=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
py=1.8.0=py37_0
py-lief=0.9.0=py37h7725739_2
pycodestyle=2.5.0=py37_0
pycosat=0.6.3=py37h14c3975_0
pycparser=2.19=py37_0
pycrypto=2.6.1=py37h14c3975_9
pycurl=7.43.0.2=py37h1ba5d50_0
pyflakes=2.1.1=py37_0
pygments=2.3.1=py37_0
pylint=2.3.1=py37_0
pyodbc=4.0.26=py37he6710b0_0
pyopenssl=19.0.0=py37_0
pyparsing=2.3.1=py37_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.14.11=py37h7b6447c_0
pysocks=1.6.8=py37_0
pytables=3.5.1=py37h71ec239_0
pytest=4.3.1=py37_0
pytest-arraydiff=0.3=py37h39e3cac_0
pytest-astropy=0.5.0=py37_0
pytest-doctestplus=0.3.0=py37_0
pytest-openfiles=0.3.2=py37_0
pytest-remotedata=0.3.1=py37_0
python=3.7.3=h0371630_0
python-dateutil=2.8.0=py37_0
python-libarchive-c=2.8=py37_6
pytz=2018.9=py37_0
pywavelets=1.0.2=py37hdd07704_0
pyyaml=5.1=py37h7b6447c_0
pyzmq=18.0.0=py37he6710b0_0
qt=5.9.7=h52cfd70_2
qtawesome=0.5.7=py37_1
qtconsole=4.4.3=py37_0
qtpy=1.7.0=py37_1
readline=7.0=h7b6447c_5
requests=2.21.0=py37_0
rope=0.12.0=py37_0
ruamel_yaml=0.15.46=py37h14c3975_0
scikit-image=0.14.2=py37he6710b0_0
scikit-learn=0.21.2=py37h627018c_0
scipy=1.3.0=py37h921218d_0
seaborn=0.9.0=py37_0
secretstorage=3.1.1=py37_0
send2trash=1.5.0=py37_0
setuptools=40.8.0=py37_0
simplegeneric=0.8.1=py37_2
singledispatch=3.4.0.3=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
snappy=1.1.7=hbae5bb6_3
snowballstemmer=1.2.1=py37_0
sortedcollections=1.1.2=py37_0
sortedcontainers=2.1.0=py37_0
soupsieve=1.8=py37_0
sphinx=1.8.5=py37_0
sphinxcontrib=1.0=py37_1
sphinxcontrib-websupport=1.1.0=py37_1
spyder=3.3.3=py37_0
spyder-kernels=0.4.2=py37_0
sqlalchemy=1.3.1=py37h7b6447c_0
sqlite=3.28.0=h8b20d00_0
statsmodels=0.9.0=py37h035aef0_0
sympy=1.3=py37_0
tblib=1.3.2=py37_0
terminado=0.8.1=py37_1
testpath=0.4.2=py37_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py37_0
tornado=6.0.2=py37h7b6447c_0
tqdm=4.31.1=py37_1
traitlets=4.3.2=py37_0
unicodecsv=0.14.1=py37_0
unixodbc=2.3.7=h14c3975_0
urllib3=1.24.1=py37_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.14.1=py37_0
wheel=0.33.1=py37_0
widgetsnbextension=3.4.2=py37_0
wrapt=1.11.1=py37h7b6447c_0
wurlitzer=1.0.2=py37_0
x264=1!152.20180806=h14c3975_0
xlrd=1.2.0=py37_0
xlsxwriter=1.1.5=py37_0
xlwt=1.3.0=py37_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.9=h516909a_1004
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.7=h14c3975_1000
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zict=0.1.4=py37_0
zipp=0.3.3=py37_1
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
